requirements.txt
asgiref==3.5.0
caldav==0.8.2
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
idna==3.3
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
TatSu==5.8.0
urllib3==1.26.9
vobject==0.9.6.1
django-bulk-update-or-create==0.3.0
